backend : ゴミ箱ディレクトリパスを"trashdir"で返すように実装
