+++ backend/tools/storage.go
@@ -1,0 +1,49 @@
+package tools
+
+import (
+	"io/ioutil"
+	"os"
+	"path/filepath"
+)
+
+type StorageItem struct {
+	Path string `json:"path"`
+	Type string `json:"type"`
+}
+
+func GetDirs(dir string) ([]string, error) {
+	findList := []string{}
+
+	items, err := ioutil.ReadDir(dir)
+
+	if err != nil {
+		return nil, err
+	}
+
+	for _, item := range items {
+		if item.IsDir() {
+			findList = append(findList, dir+"/"+item.Name())
+		}
+	}
+
+	return findList, err
+}
+
+// return dir and file of paths and types
+func GetDirAndFilePaths(dir string) ([]StorageItem, error) {
+	files, err := os.ReadDir(dir)
+	if err != nil {
+		return nil, err
+	}
+
+	var items []StorageItem
+	for _, file := range files {
+		if file.IsDir() {
+			items = append(items, StorageItem{Path: filepath.Join(dir, file.Name()), Type: "dir"})
+			continue
+		}
+		items = append(items, StorageItem{Path: filepath.Join(dir, file.Name()), Type: "file"})
+	}
+
+	return items, nil
+}
